Fix broken MongoidAdapter expectation
Revert "Refactor extract TokenAuthenticatable"
This reverts commit e797557c3d54cc15bf248e04431121164359e242.

The MongoidAdapter is built on the expectation that its `base_class`
will include SimpleTokenAuthentication::ActsAsTokenAuthenticatable
and will automatically provide .acts_as_token_authenticatable to
the classes which will include it.

Example:

    class User
      include Mongoid::Document # the MongoidAdapter.base_class

      # is expected to be able to use
      acts_as_token_athenticatable
    end

That expectation was broken in e797557 because ActsAsTokenAuthenticatable
module did no longer include ActiveSupport::Concern, hence lost its
ability to define class methods when `include`d (vs. `extend`ed).

For the purpose of defining class methods, including two successive
concerns does the job, while including a concern (Mongoid::Document)
that extended ActsAsTokenAuthenticatable (no longer a concern) does not.

Many thanks @krsyoung for pinpointing the cause of the bug.

Fixes #168
