+++ lib/simple_token_authentication.rb
@@ -11,7 +11,7 @@ module SimpleTokenAuthentication
 
   def self.ensure_models_can_act_as_token_authenticatables model_adapters
     model_adapters.each do |model_adapter|
-      model_adapter.base_class.send :extend, SimpleTokenAuthentication::ActsAsTokenAuthenticatable
+      model_adapter.base_class.send :include, SimpleTokenAuthentication::ActsAsTokenAuthenticatable
     end
   end
